Separated types from data and queries
